+++ app/templates/error.html
@@ -1,0 +1,6 @@
+{{ define "content" }}
+
+<h1>Error</h1>
+<h3>{{ . }}</h3>
+
+{{ end }}
